+++ assets/Flot.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace dimple\administrator\assets;
+
+use yii\web\AssetBundle;
+
+class Flot extends AssetBundle
+{
+    public $sourcePath = '@bower/flot';
+    public $js = [
+        'jquery.flot.js'
+    ];
+    public $depends = [
+        'yii\web\JqueryAsset'
+    ];
+}