+++ controllers/SystemInformationController.php
@@ -14,26 +14,31 @@ class SystemInformationController extends Controller
 {
     public function actionIndex()
     {
+        $provider = \probe\Factory::create();
+        echo $provider->getCpuModel();
+        $cpu =  $provider->getCpuUsage();
+        print_r($cpu);
+        echo $provider->getFreeMem();
 
-        $provider = Factory::create();
-        if ($provider) {
-            if (Yii::$app->request->isAjax) {
-                Yii::$app->response->format = Response::FORMAT_JSON;
-                if ($key = Yii::$app->request->get('data')) {
-                    switch($key){
-                        case 'cpu_usage':
-                            return$provider->getCpuUsage();
-                            break;
-                        case 'memory_usage':
-                            return ($provider->getTotalMem() - $provider->getFreeMem()) / $provider->getTotalMem();
-                            break;
-                    }
-                }
-            } else {
-                return $this->render('index', ['provider' => $provider]);
-            }
-        } else {
-            return $this->render('fail');
-        }
+
+        // if ($provider) {
+        //     if (Yii::$app->request->isAjax) {
+        //         Yii::$app->response->format = Response::FORMAT_JSON;
+        //         if ($key = Yii::$app->request->get('data')) {
+        //             switch($key){
+        //                 case 'cpu_usage':
+        //                     return$provider->getCpuUsage();
+        //                     break;
+        //                 case 'memory_usage':
+        //                     return ($provider->getTotalMem() - $provider->getFreeMem()) / $provider->getTotalMem();
+        //                     break;
+        //             }
+        //         }
+        //     } else {
+        //         return $this->render('index', ['provider' => $provider]);
+        //     }
+        // } else {
+        //     return $this->render('fail');
+        // }
     }
 }