+++ CHANGELOG.md
@@ -16,4 +16,5 @@ Features:
  - create, update, delete, reset password
  - block account
  - confirmed account
- 
+- user log monitoring
+- system log monitroing